191 Add ability to add assets
- add ability to create new assets.
- update date time component to support editing / adding assets.
- seperate add item form and add item instructions
- add ability to remove selected inventory items.
- tech debt and minor bug fixes

Todo:
1. support bulk upload
