+++ frontend/src/features/InventoryList/AddInventory/AddInventoryInstructions.jsx
@@ -1,0 +1,35 @@
+import { Stack, Typography } from '@mui/material';
+
+export default function AddInventoryInstructions({ stepNumber }) {
+  switch (stepNumber) {
+    case 1:
+      return (
+        <Stack paddingBottom="2rem">
+          <Typography> Fill in the necessary details.</Typography>
+          <Typography variant="caption">
+            These details will help us quickly search for the items that you are looking for later on.
+          </Typography>
+        </Stack>
+      );
+    case 2:
+      return (
+        <Stack paddingBottom="2rem">
+          <Typography> Fill in the optional details.</Typography>
+          <Typography variant="caption">
+            Extra details helps you understand your product limitations, expiry dates and much more.
+          </Typography>
+        </Stack>
+      );
+    case 3:
+      return (
+        <Stack paddingBottom="2rem">
+          <Typography> Confirm your changes.</Typography>
+          <Typography variant="caption">
+            Adding new items as assets with product details. You can always edit your product later on.
+          </Typography>
+        </Stack>
+      );
+    default:
+      return null;
+  }
+}
