+++ frontend/src/features/InventoryList/EditInventory.jsx
@@ -105,6 +105,7 @@ const EditInventory = () => {
 
   useEffect(() => {
     if (id.length > 0) {
+      console.log(id);
       dispatch(inventoryActions.getInvByID(id));
     }
   }, [id]);
